% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{boxplot_mir}
\alias{boxplot_mir}
\title{Boxplot of individual miRNA count compared by the first condition}
\usage{
boxplot_mir(
  dds,
  mir_of_interest,
  condition_list = "conditions",
  title = "Project",
  y_label = "normalised counts - log scale"
)
}
\arguments{
\item{dds}{DESeq dataset}

\item{mir_of_interest}{List of mir to plot, ideally 25 or lesss}

\item{condition_list}{a string vector describing the different conditions}

\item{title}{title to add to the plot, default : "Project"}

\item{y_label}{string to add as a label on the y axis}
}
\description{
Boxplot of individual miRNA count compared by the first condition
}
