% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{boxplot_one_mir}
\alias{boxplot_one_mir}
\title{Boxplot of individual miRNA count compared by the first condition}
\usage{
boxplot_one_mir(
  dds,
  mir,
  title = NA,
  condition_list = "conditions",
  y_label = "normalised counts - log scale"
)
}
\arguments{
\item{dds}{DESeq dataset}

\item{mir}{name of a single mir to plot (eg : "omy-miR-375-3p")}

\item{title}{title to add to the plot, default is the mir name}

\item{condition_list}{a string vector describing the different conditions}

\item{y_label}{string to add as a label on the y axis}
}
\description{
Boxplot of individual miRNA count compared by the first condition
}
