% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{compute_all_resampling_combinaisons}
\alias{compute_all_resampling_combinaisons}
\title{Compute all the resampling combinaisons to subset biosamples}
\usage{
compute_all_resampling_combinaisons(
  sample_info,
  condition_name = "conditions",
  subset_size = 4,
  subset_proportion = 0.66
)
}
\arguments{
\item{sample_info}{sample info tibble with a conditions column}

\item{condition_name}{a string describing the name of the condition which is a column of sample_info}

\item{subset_size}{the size of the subsets to resample, if subset_proportion is set to NA}

\item{subset_proportion}{the proportion of the condition to resample, if set to NA the size will be constant at subset_size}
}
\value{
list of all subset_size combinaison for each condition
}
\description{
Compute all the resampling combinaisons to subset biosamples
}
