% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{compute_resampling}
\alias{compute_resampling}
\title{Run the analysis function for n_resample random subset of biosamples}
\usage{
compute_resampling(
  reads,
  sample_info,
  condition_list,
  ds_formula,
  analysis_function,
  control_combinaisons,
  treatment_combinaisons,
  n_resample = 50
)
}
\arguments{
\item{reads}{reads tibble}

\item{sample_info}{sample info tibble with a conditions column}

\item{condition_list}{a string vector describing the different conditions}

\item{ds_formula}{a formula describing the design of the analysis}

\item{analysis_function}{custom analysis function using the reads, sample_info, condition_list and ds_formula}

\item{control_combinaisons}{a tibble with control biosample subsets to analyse 
(a line is a subset, a column is a biosample, column names doesnt matter)}

\item{treatment_combinaisons}{a tibble with treatment biosample subsets to analyse 
(a line is a subset, a column is a biosample, column names doesnt matter)}

\item{n_resample}{the number of resampling to analyse}
}
\description{
Run the analysis function for n_resample random subset of biosamples
}
