% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cpm_distribution_histogram}
\alias{cpm_distribution_histogram}
\title{Histogram of the count distribution}
\usage{
cpm_distribution_histogram(
  compressed_by_annotation,
  normalized = T,
  xlim = c(0, 100),
  title = "Project"
)
}
\arguments{
\item{compressed_by_annotation}{compressed_by_annotation tibble}

\item{normalized}{whether to use raw or normalized from prost, default normalized}

\item{xlim}{A numeric vector of length two providing limits of the scale}

\item{title}{title to add to the plot, default : "Project"}
}
\value{
ggplot histogram
}
\description{
The range of value is from 0 to 100,
}
