% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{dds_from_compressed_by_annotation}
\alias{dds_from_compressed_by_annotation}
\title{Convert the compressed_by_annotation tibble into a DESeq dataset}
\usage{
dds_from_compressed_by_annotation(
  compressed_by_annotation,
  sample_info,
  condition_list = "conditions",
  ds_formula = ~conditions
)
}
\arguments{
\item{compressed_by_annotation}{compressed_by_annotation tibble}

\item{sample_info}{sample info tibble with a condition column}

\item{condition_list}{a string vector describing the different conditions}

\item{ds_formula}{a formula describing the design of the analysis}
}
\value{
DESeq dataset
}
\description{
Convert the compressed_by_annotation tibble into a DESeq dataset
}
