% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prost_data.R
\name{load_metadata}
\alias{load_metadata}
\title{Load metadata from a sample info csv file}
\usage{
load_metadata(path, condition_list = "condition")
}
\arguments{
\item{path}{path of the sample info file}

\item{condition_list}{a string vector describing the different conditions}
}
\value{
tibble containing the csv file with an extra sample_ID column
}
\description{
Load metadata from a sample info csv file
}
\examples{
\dontrun{
load_metadata("sample_info/eggpreserve.csv")
}

}
