% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prost_data.R
\name{load_prost_compressed_by_annotation}
\alias{load_prost_compressed_by_annotation}
\title{Load data from the prost compressed_by_annotation output file}
\usage{
load_prost_compressed_by_annotation(path)
}
\arguments{
\item{path}{path of the data path}
}
\value{
tibble containing the raw and normalized count by annotated miRNA of each sample
}
\description{
When read counts for multiple annotations, the counts are split equaly between
each annotation
}
\examples{
\dontrun{
load_prost_compressed_by_annotation("prost_11/eggpreserve_compressed_by_annotation.tsv")
}

}
