% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prost_data.R
\name{load_prost_compressed_by_genomic_location_miRNA_only}
\alias{load_prost_compressed_by_genomic_location_miRNA_only}
\title{Load data from the prost compressed_by_location output file}
\usage{
load_prost_compressed_by_genomic_location_miRNA_only(path)
}
\arguments{
\item{path}{path to the file}
}
\description{
Load data from the prost compressed_by_location output file
}
