% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prost_data.R
\name{mir_in_every_condition}
\alias{mir_in_every_condition}
\title{List the miRNAs counted in every condition}
\usage{
mir_in_every_condition(count_per_condition, threshold = 10)
}
\arguments{
\item{count_per_condition}{count_per_condition tibble}

\item{threshold}{min threshold to be considered as present}
}
\value{
a single column tibble with the list of miRNAs
}
\description{
List the miRNAs counted in every condition
}
