% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prost_data.R
\name{mir_locations}
\alias{mir_locations}
\title{Get location information for specific mirs}
\usage{
mir_locations(compressed_by_genomic_location, sample_info, mir_list)
}
\arguments{
\item{compressed_by_genomic_location}{compressed_by_genomic_location tibble}

\item{sample_info}{sample_info tibble}

\item{mir_list}{vector of micro RNA to be searched}
}
\description{
Get location information for specific mirs
}
