% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pca_plot}
\alias{pca_plot}
\title{PCA plot of the samples}
\usage{
pca_plot(
  deseq_transformation,
  condition_list = "conditions",
  title = "Project",
  add_theme = T,
  draw_arrows = T,
  n_arrows = 10,
  draw_ellipses = T,
  draw_sample_label = F
)
}
\arguments{
\item{deseq_transformation}{DESeq transformation (e.g. normTransform(dds))}

\item{condition_list}{a string vector describing the different conditions}

\item{title}{title to add to the plot, default : "Project"}

\item{add_theme}{if TRUE then a theme is added to the figure, labels, title, colors...}

\item{draw_arrows}{if TRUE then miR contribution arrows are added to the figure}

\item{n_arrows}{number of arrow to add to the figure, up to 10}

\item{draw_ellipses}{if TRUE then ellipses around conditions are added to the figure}

\item{draw_sample_label}{if TRUE then sample names are added to the figure}
}
\value{
ggplot point
}
\description{
PCA plot of the samples
}
