% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mir_origin_from_juanchich_data}
\alias{plot_mir_origin_from_juanchich_data}
\title{Plot level of expression in every organ of omy miRNA}
\usage{
plot_mir_origin_from_juanchich_data(
  omy_miRNA,
  plot_title,
  path_MiRNAOrigin = "~/Documents/phenomir/askomics/datatables/MiRNAOrigin.csv",
  path_MiRNA = "~/Documents/phenomir/askomics/datatables/MiRNA.csv",
  path_Organ = "~/Documents/phenomir/askomics/datatables/Organ.csv"
)
}
\arguments{
\item{omy_miRNA}{list of miRNA patterns that will be interpreted as a regex}

\item{plot_title}{title of the plot}

\item{path_MiRNAOrigin}{data path of a MiRNAOrigin askomics table}

\item{path_MiRNA}{data path of a MiRNA askomics table}

\item{path_Organ}{data path of a Organ askomics table}
}
\description{
Plot level of expression in every organ of omy miRNA
}
