% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prost_data.R
\name{raw_read_count_per_sample}
\alias{raw_read_count_per_sample}
\title{Sum the total read counts for each sample}
\usage{
raw_read_count_per_sample(compressed_by_annotation)
}
\arguments{
\item{compressed_by_annotation}{compressed_by_annotation tibble}
}
\value{
a tibble with a column sample and a column total_count
}
\description{
Sum the total read counts for each sample
}
