% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prost_data.R
\name{read_count_per_condition}
\alias{read_count_per_condition}
\title{Sum the read counts for every sample by condition}
\usage{
read_count_per_condition(
  compressed_by_annotation,
  sample_info,
  condition_list = "conditions",
  normalized = F
)
}
\arguments{
\item{compressed_by_annotation}{compressed_by_annotation tibble}

\item{sample_info}{sample_info tibble}

\item{condition_list}{a string vector describing the different conditions}

\item{normalized}{whether to use raw or normalized from prost, default raw}
}
\value{
a tibble with a line by condition and a column by annotated miRNA
}
\description{
The condition is from the conditions column in sample_info
}
