% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{resampling_barplot}
\alias{resampling_barplot}
\title{Barplot describing resampling analysis, for each miR found differential at least one time the number of time it was found differential}
\usage{
resampling_barplot(
  reference_analysis,
  resampling_results,
  n_resample = 50,
  threshold_padj = 0.05,
  threshold_lfc = 0.58,
  title = ""
)
}
\arguments{
\item{reference_analysis}{tibble from compute_resampling function}

\item{resampling_results}{analysis whit every biosamples}

\item{n_resample}{the number of resampling to analyse}

\item{threshold_padj}{maximum adjusted pvalue to be considered differentially expressed (default 0.05)}

\item{threshold_lfc}{minimum log2 fold change to be considered differentially expressed (default 0.58, corresponds to 1.5 fold change)}

\item{title}{}
}
\description{
Barplot describing resampling analysis, for each miR found differential at least one time the number of time it was found differential
}
