% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{resampling_heatmap}
\alias{resampling_heatmap}
\title{Heatmap describing resampling analysis}
\usage{
resampling_heatmap(
  sample_info,
  resampling_results,
  subset_size = 4,
  threshold_padj = 0.05,
  threshold_lfc = 0.58,
  title = "",
  condition_name = "conditions"
)
}
\arguments{
\item{sample_info}{sample_info tibble}

\item{resampling_results}{analysis whit every biosamples}

\item{subset_size}{number of sample for each condition in a resampling}

\item{condition_name}{condition used to facet the heatmap}
}
\description{
Heatmap describing resampling analysis
}
