% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sample_to_mir_heatmap}
\alias{sample_to_mir_heatmap}
\title{Heatmap sample to miRMA}
\usage{
sample_to_mir_heatmap(
  dds,
  deseq_transformation,
  cluster_cols = T,
  condition_list = "conditions",
  scale = "row",
  title = "Project"
)
}
\arguments{
\item{dds}{Normalized DESeq dataset}

\item{deseq_transformation}{DESeq transformation (e.g. normTransform(dds))}

\item{cluster_cols}{boolean values determining if columns should be clustered or hclust object}

\item{condition_list}{a string vector describing the different conditions}

\item{scale}{character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. Corresponding values are "row", "column" and "none"}

\item{title}{title to add to the plot, default : "Project"}
}
\value{
pheatmap heatmap
}
\description{
Warning : requires metadata to have "condition" and "project_name" entries
}
