% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sample_to_sample_heatmap}
\alias{sample_to_sample_heatmap}
\title{Heatmap sample to sample}
\usage{
sample_to_sample_heatmap(dds, deseq_transformation, title = "Project")
}
\arguments{
\item{dds}{Normalised DESeq dataset}

\item{deseq_transformation}{DESeq transformation (e.g. normTransform(dds))}

\item{title}{title to add to the plot, default : "Project"}
}
\value{
pheatmap heatmap
}
\description{
Heatmap sample to sample
}
