% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prost_data.R
\name{select_mir_compressed_by_annotation}
\alias{select_mir_compressed_by_annotation}
\title{Subset the compressed_by_annotation tibble to keep only mir with a minimum rpm}
\usage{
select_mir_compressed_by_annotation(
  compressed_by_annotation,
  sample_info,
  condition_col = "condition",
  min_rpm = 10,
  min_total_count = 0,
  min_prop = 0.7
)
}
\arguments{
\item{compressed_by_annotation}{compressed_by_annotation tibble}

\item{sample_info}{sample_info tibble}

\item{condition_col}{string, name of the condition column in sample_info}

\item{min_rpm}{numeric, minimum rpm required for at least some samples}

\item{min_total_count}{numeric, minimum total count required}

\item{min_prop}{numeric, minimum proportion of samples in the smallest group}
}
\description{
Subset the compressed_by_annotation tibble to keep only mir with a minimum rpm
}
