% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prost_data.R
\name{subset_samples_compressed_by_annotation}
\alias{subset_samples_compressed_by_annotation}
\title{Subset the compressed_by_annotation tibble with a sample list}
\usage{
subset_samples_compressed_by_annotation(compressed_by_annotation, sample_list)
}
\arguments{
\item{compressed_by_annotation}{compressed_by_annotation tibble}

\item{sample_list}{list of sample to keep}
}
\description{
Subset the compressed_by_annotation tibble with a sample list
}
