% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_prost_data.R
\name{subset_samples_sample_info}
\alias{subset_samples_sample_info}
\title{Subset the sample_info tibble with a sample list}
\usage{
subset_samples_sample_info(sample_info, sample_list)
}
\arguments{
\item{sample_info}{sample_info tibble}

\item{sample_list}{list of sample to keep}
}
\description{
Subset the sample_info tibble with a sample list
}
